30-07-18/project-20180728T172649Z-001/project/src/basketballProject/Video.java
package basketballProject;

import org.json.JSONObject;

public class Video {
	private int videoID;
	private String videoTitle;
	private String videoURL;
	private DBC db = new DBC();
	private JSONObject jsonObj;
	
	public Video(String videoTitle, String videoURL){
		videoID = db.getNewestID("Video_ID", "video") + 1;
		this.videoTitle = videoTitle;
		this.videoURL = videoURL;
	}
	
	public int getID(){
		return videoID;
	}
	public String getVideoTitle(){
		return videoTitle;
	}
	
	public String videoURL(){
		return videoURL;
	}
	
	public JSONObject toJson() {
		jsonObj = new JSONObject();
		jsonObj.put("videoTitle", videoTitle);
		jsonObj.put("videoURL", videoURL);
		return jsonObj;
	}
}
